select contest checkboxes
